More PR comments addressed
