--- Sources/TuistCache/Cache/CacheAPIError.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-enum CacheAPIError: LocalizedError {
-    case incorrectCloudURL
-    case missingCloudConfig
-}
